Update solution for expected-value-discrete

+++ expected-value-discrete/expected-value-discrete.py
@@ -7,8 +7,8 @@ def expected_value_discrete(x, p):
     # Write code here
     # step 0
     # 先将输入转换为NumPy数组，并指定高精度浮点类型
-    x_arr = np.array(x, dtype=np.float64)
-    p_arr = np.array(p, dtype=np.float64)
+    x_arr = np.array(x)
+    p_arr = np.array(p)
     
     # 验证概率和（使用更宽松的容差，如1e-6）
     if not np.allclose(np.sum(p_arr), 1.0, atol=1e-6):
